requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
boto3==1.12.24
botocore==1.15.24
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.3
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
isodate==0.6.0
jmespath==0.9.5
Pillow==7.0.0
psycopg2==2.8.4
PyLyrics==1.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
numpy==1.20.2
opencv-python==4.5.1.48
youtube-dl==2021.4.7
python-vlc
pafy
